[KT] Esimd radix sort: key-value (#1208)
* [KT] Explicitly pass memory used by onesweep kernel

* [KT] key-value sort
---------
